fix(build): relocate components path
