+++ src/renderer/src/components/settings/ModelConfigItem.vue
